==> 03_evolutionAsylum.Rmd <==
---
title: ""
author: "Duc-Quang Nguyen | swissinfo.ch"
date: " 2017"
output: 
  html_document:
    toc: true
    toc_depth: 3
    theme: simplex
---

```{r setup, include=FALSE}
data.file.full <- "data/yearlyAsylum_eurostat_full.csv"
data.byContinent <- "input/yearlyAsylum_byContinent.csv"
data.tot <- "input/yearlyAsylum_total.csv"

library(readr)
library(tidyr)
library(dplyr)
library(magrittr)
library(countrycode)
library(ggplot2)
library(scales)
library(swiMap)
library(swiTheme)
library(forcats)

### Getting data in packages
library(eurostat)

```

```{r load data}
data.read <- read.csv(data.file.full) %>% 
  filter(sex == "Total", age == "Total", asyl_app == "First time applicant") %>%
  select(-sex, -age, -asyl_app, -unit)

```

```{r evolution asylum seeker}
iso2agg <- c("EU28", "TOTAL")
citizenAgg <- c("Total", "European Union (28 countries)", "Extra-EU-28")

# yearly total check
tot.yearly <- data.read %>% filter(iso2 == "TOTAL", citizen == "Total") %>%
  select(time, values) %>% arrange(time)

dat <- data.read %>% filter(!iso2 %in% iso2agg, !citizen %in% citizenAgg)

# get the contient of origin
dat$citizen <- as.character(dat$citizen)
citizen <- structure(countrycode(unique(dat$citizen), "country.name", "continent"), 
          names = unique(as.character(dat$citizen)))
#
citizen[which(names(citizen) == 'Kosovo (under United Nations Security Council Resolution 1244/99)')] <- 'Europe'
dat$cit.cont <- citizen[match(dat$citizen, names(citizen))]

dat %<>% group_by(cit.cont, time) %>% dplyr::summarise(tot = sum(values, na.rm = T )) %>% 
  ungroup() %>% arrange(time)

## check
tot.yearly2 <- dat %>% group_by(time) %>% dplyr::summarise(sumVal = sum(tot, na.rm = T )) %>% ungroup
tot.check <- left_join(tot.yearly, tot.yearly2)
stopifnot(abs(tot.check$values - tot.check$sumVal)<1000)

dat %<>% spread(cit.cont, tot)
dat$others <- dat$Oceania +  dat$`<NA>`
dat %<>% select(-Oceania, -`<NA>`)

write.csv(dat, data.byContinent, row.names = F)
write.csv(tot.yearly, data.tot, row.names = F)
```


```{}
```
